+++ src/main/scala/org/economicsl/auctions/actors/AuctionActor.scala
@@ -3,6 +3,7 @@ package org.economicsl.auctions.actors
 import akka.actor.{ActorRef, Terminated}
 import akka.routing.{ActorRefRoutee, BroadcastRoutingLogic, Router}
 import org.economicsl.auctions._
+import org.economicsl.auctions.participants.Token
 import org.economicsl.core.Tradable
