+++ src/main/scala/org/economicsl/auctions/participants/AuctionParticipant.scala
@@ -13,8 +13,9 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package org.economicsl.auctions
+package org.economicsl.auctions.participants
 
+import org.economicsl.auctions.Auction
 import org.economicsl.core.Tradable
