+++ src/main/scala/org/economicsl/auctions/participants/OrderIssuer.scala
@@ -13,16 +13,14 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package org.economicsl.auctions
+package org.economicsl.auctions.participants
 
-import org.economicsl.core.util.UUIDGenerator
+import org.economicsl.auctions.Issuer
 
+trait OrderIssuer[A <: OrderIssuer[A]]
+  extends TokenGenerator {
+  this: A =>
 
-trait TokenGenerator
-  extends UUIDGenerator {
-
-  protected def randomToken(): Token = {
-    random()
-  }
+  def issuer: Issuer
 
 }
