+++ src/test/scala/org/economicsl/auctions/participants/TestAuctionParticipant.scala
@@ -13,7 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package org.economicsl.auctions.singleunit
+package org.economicsl.auctions.participants
 
 import org.economicsl.auctions.Auction.AuctionProtocol
 import org.economicsl.auctions._
